pyproject.toml
[project]
name = "battery-agent"
version = "0.1.0"
description = "LLM-Based Battery Arbitrage Agent with Reflexion - 基于 LangGraph 的智能体系统，模拟家庭储能电池在分时电价市场中的套利行为"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
battery-agent = "main:main"

